requirements.txt
PyQt6>=6.5.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
hypothesis>=6.82.0
pytest>=7.4.0
pytest-qt>=4.2.0
pyinstaller>=6.0.0
